Move REST-API interfaces to separate package
