+++ servers/services/src/main/java/org/projectnessie/services/rest/ContentsResource.java
@@ -20,7 +20,7 @@ import java.security.Principal;
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
-import org.projectnessie.api.ContentsRestApi;
+import org.projectnessie.api.rest.ContentsRestApi;
 import org.projectnessie.error.NessieConflictException;
 import org.projectnessie.error.NessieNotFoundException;
 import org.projectnessie.model.CommitMeta;
