+++ servers/services/src/main/java/org/projectnessie/services/rest/TreeResource.java
@@ -21,7 +21,7 @@ import java.util.List;
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
-import org.projectnessie.api.TreeRestApi;
+import org.projectnessie.api.rest.TreeRestApi;
 import org.projectnessie.error.NessieConflictException;
 import org.projectnessie.error.NessieNotFoundException;
 import org.projectnessie.model.Branch;
